fixes: #24 pr number via github.event.number (#26)
* fixes: #24 pr number via github.event.number

* ci: poor guy (#27)

* ci: poor guy

* ci: fix npm command

* ci: that bad? ok

* spam: no thanks (#29)

* spam: no thanks

* Using one block of this action was causing manu API calls to Github API (#31)

* Using one block of this action was causing manu API calls to Github API
This is because in the composite action it uses tj-actions/branch-names and peter-evans/create-or-update-comment actions as uses
Contrary, even when there are if: conditions to not execute those actions, API calls to Git are still made to obtain those actions meta data
Further, if we support the issue #25, it will exponentially add more API calls. More uses: more api calls to Github, regardless of the if:
Solved it by implement whole in whole via shell.
Now this action will only make one API call when used. Just the first time run it makes a secondary API call on the backend which is further cached for respective calls.

* ui: a few fixes

* ci: unknown file